refactor(WIP): steam auth
